Move report related types to own module
Reduces risk of circular imports
